+++ src/Oro/Bundle/TestFrameworkBundle/BehatStatisticExtension/AvgTimeProvider/MasterAvgTimeProvider.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Oro\Bundle\TestFrameworkBundle\BehatStatisticExtension\AvgTimeProvider;
+
+class MasterAvgTimeProvider extends AbstractAvgTimeProvider
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected function calculate()
+    {
+        $criteria = [
+            'git_branch' => 'master',
+        ];
+
+        $this->averageTimeTable = $this->repository->getAverageTimeTable($criteria);
+        $this->calculateAverageTime();
+    }
+}